add hashmap

+++ libft/srcs/hash_map/hash_ascii.c
@@ -6,7 +6,7 @@
 /*   By: tkobb <tkobb@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/22 00:21:55 by tkobb             #+#    #+#             */
-/*   Updated: 2018/10/22 19:48:12 by tkobb            ###   ########.fr       */
+/*   Updated: 2018/10/24 22:57:04 by tkobb            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 int		hash_ascii(const char *str)
 {
-	int		hash;
+	size_t	hash;
 	int		len;
 	int		i;
